add: Authorization

+++ routers/posts.py
@@ -8,7 +8,7 @@ from sqlalchemy.orm import joinedload
 import models
 from database import get_db
 from schemas import PostCreate, PostResponse, PostUpdate
-
+from auth import CurrentUser
 
 router = APIRouter()
 
@@ -32,18 +32,13 @@ async def get_posts(db: Annotated[AsyncSession, Depends(get_db)]):
 )
 async def create_post(
     post: PostCreate,
+    current_user:CurrentUser,
     db: Annotated[AsyncSession, Depends(get_db)]
 ):
-    user = await db.get(models.User, post.user_id)
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="User not found",
-        )
     new_post = models.Post(
         title=post.title,
         content=post.content,
-        user_id=post.user_id,
+        user_id=current_user.id,
     )
     db.add(new_post)
     await db.commit()
@@ -75,6 +70,7 @@ async def get_post(
 async def update_post_full(
         post_id: int,
         post_date: PostCreate,
+        current_user:CurrentUser,
         db: Annotated[AsyncSession, Depends(get_db)],
 ):
     stmt = (
@@ -89,16 +85,15 @@ async def update_post_full(
             status_code=status.HTTP_404_NOT_FOUND, 
             detail="Post not found"
         )
-    if post_date.user_id != post.user_id:
-        new_user = await db.get(models.User, post_date.user_id)
-        if not new_user:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail="User not found",
-            )
+        
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to update this post",
+        )
+        
     post.title = post_date.title
     post.content = post_date.content
-    post.user_id = post_date.user_id
     
     await db.commit()
     await db.refresh(post, attribute_names=["author"])
@@ -109,6 +104,7 @@ async def update_post_full(
 async def update_post_partial(
     post_id: int,
     post_data: PostUpdate,
+    current_user: CurrentUser,
     db: Annotated[AsyncSession, Depends(get_db)],
 ):
     stmt = (
@@ -120,7 +116,14 @@ async def update_post_partial(
     post = result.scalars().first()
     if not post:
         raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="Post not found"
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="Post not found"
+        )
+        
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to update this post",
         )
 
     update_data = post_data.model_dump(exclude_unset=True)
@@ -136,12 +139,20 @@ async def update_post_partial(
 @router.delete("/{post_id}", status_code=status.HTTP_204_NO_CONTENT)
 async def delete_post(
     post_id: int,
+    current_user: CurrentUser,
     db: Annotated[AsyncSession, Depends(get_db)]
 ):
     post = await db.get(models.Post, post_id)
     if not post:
         raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="Post not found"
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="Post not found"
+        )
+        
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to delete this post",
         )
 
     await db.delete(post)
